Проведен рефакторинг, вынесены общие методы для стратегий

+++ Data.h
@@ -5,9 +5,9 @@
 struct Data
 {
     QString m_name;
-    int m_size;
+    qint64 m_size;
     qreal m_ratio;
-    Data(const QString& name, int size, qreal ratio)
+    Data(const QString& name, qint64 size, qreal ratio)
                 : m_name(name), m_size(size), m_ratio(ratio) {}
 };
 #endif // DATA_H
